split up routes from main & add series form
can convert from the form to an insertable model that has the
validation.
added error chain for sane erorr handling.
the input from the series form is validated & dumps what
would be inserted into the database as a json response,
this is a test of how ergonomic this style is with
form->insertable->validate->insert